Add index.html
There are 2 images. One of them is a regular image while another image
which is an SVG built using vector paths is superimposed on it. The
superimposed image is of the exact object as in the background image
except that it is void of any background and is placed at the exact
position where the object is in background image. The
user's goal is to click on the object. Different eventListeners are
added to both images.