Update login and register pages with auth.css
This change make the error message shakes so that it is easier for users
to percieve that there is an error. Earlier, users would get same error
message multiple times and would think that the website has stopped
responding. But now, the error message will shake so the website will
look responsive. This is done by using animation in auth.css and use key
in the error component which is updated each time the error comes again
by incrementing the message key so that react rerender it each time and
the animation is shown each time.