Add isl_converter.py and corresponding endpoint to main.py
A new feature is being made to convert english sentences to their
english representation of sequence of signs in Indian sign language. In
simple words, it should translate engilsh sentences to a sequence of
words whose corresponding signs should be done in ISL. The logic for
this is in isl_converter.py and the endpoint is made in main.py. This
feature is not completed but is integrated to analyze its performance.